1.0: submodule djvulibre
