+++ src/jb2cmp/README
