+++ src/normalize.h
@@ -4,30 +4,30 @@
  * Copyright 2012, Mikhail Dektyarev <mihail.dektyarow@gmail.com>
  *
  * This file is part of Simpledjvu.
- * 
+ *
  * Simpledjvu is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
- * 
+ *
  * Simpledjvu is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public License
  * along with Simpledjvu.  If not, see <http://www.gnu.org/licenses/>.
  *
  */
 
-#ifndef TYPES_H_
-#define TYPES_H_
+#ifndef NORMALIZE_H_
+#define NORMALIZE_H_
 
-typedef int int32;
-typedef unsigned char byte;
+#include <djvulibre.h>
 
-const int COLORS_COUNT = 256;
-const byte MIN_COLOR = 0;
-const byte MAX_COLOR = COLORS_COUNT - 1;
+#include <types.h>
 
-#endif  // TYPES_H_
+GP<GBitmap> get_norm_image(const GBitmap &image, const int iterations = 20);
+void rescale_bitmap(const GBitmap &in, GBitmap &out);
+
+#endif  // NORMALIZE_H_
