+++ src/types.h
@@ -4,33 +4,30 @@
  * Copyright 2012, Mikhail Dektyarev <mihail.dektyarow@gmail.com>
  *
  * This file is part of Simpledjvu.
- * 
+ *
  * Simpledjvu is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
- * 
+ *
  * Simpledjvu is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public License
  * along with Simpledjvu.  If not, see <http://www.gnu.org/licenses/>.
  *
  */
 
-#ifndef DJVULIBRE_H_
-#define DJVULIBRE_H_
+#ifndef TYPES_H_
+#define TYPES_H_
 
-#include <DjVuGlobal.h>
-#include <GSmartPointer.h>
-#include <IFFByteStream.h>
-#include <GBitmap.h>
-#include <JB2Image.h>
-#include <DjVuInfo.h>
-#include <GURL.h>
-#include <GScaler.h>
-#include <IW44Image.h>
+typedef int int32;
+typedef unsigned char byte;
 
-#endif  // DJVULIBRE_H_
+const int COLORS_COUNT = 256;
+const byte MIN_COLOR = 0;
+const byte MAX_COLOR = COLORS_COUNT - 1;
+
+#endif  // TYPES_H_
